+++ source/Nuke.Common/Tools/InspectCode/InspectCode.Generated.cs
@@ -28,11 +28,11 @@ namespace Nuke.Common.Tools.InspectCode
         /// <summary><p>Path to the InspectCode executable.</p></summary>
         public static string InspectCodePath => ToolPathResolver.GetPackageExecutable("JetBrains.ReSharper.CommandLineTools", GetPackageExecutable());
         /// <summary><p>One of ReSharper's most notable features, code inspection, is available even without opening Visual Studio. InspectCode, a free command line tool requires a minimum of one parameter- your solution file- to apply all of ReSharper's inspections.</p></summary>
-        public static IEnumerable<string> InspectCode(string arguments, string workingDirectory = null, ProcessSettings processSettings = null)
+        public static IEnumerable<string> InspectCode(string arguments, string workingDirectory = null, IReadOnlyDictionary<string, string> environmentVariables = null, int? timeout = null, bool redirectOutput = false, Func<string, string> outputFilter = null)
         {
-            var process = ProcessTasks.StartProcess(InspectCodePath, arguments, workingDirectory, processSettings?.EnvironmentVariables, processSettings?.ExecutionTimeout, processSettings?.RedirectOutput ?? true);
+            var process = ProcessTasks.StartProcess(InspectCodePath, arguments, workingDirectory, environmentVariables, timeout, redirectOutput, outputFilter);
             process.AssertZeroExitCode();
-            return process.Output.Select(x => x.Text);
+            return process.HasOutput ? process.Output.Select(x => x.Text) : null;
         }
         static partial void PreProcess(InspectCodeSettings toolSettings);
         static partial void PostProcess(InspectCodeSettings toolSettings);