+++ source/Nuke.Common/Utilities/Collections/Enumerable.SingleOrDefaultOrError.cs
@@ -1,17 +1,19 @@
-﻿using System;
+﻿// Copyright Matthias Koch, Sebastian Karasek 2018.
+// Distributed under the MIT License.
+// https://github.com/nuke-build/nuke/blob/master/LICENSE
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace Nuke.Common.Utilities.Collections
 {
     public static partial class EnumerableExtensions
     {
+        [CanBeNull]
         public static T SingleOrDefaultOrError<T>(this IEnumerable<T> enumerable, Func<T, bool> predicate, string message)
         {
-            if (enumerable == null) throw new ArgumentNullException(nameof(enumerable));
-            if (predicate == null) throw new ArgumentNullException(nameof(predicate));
-            if (string.IsNullOrEmpty(message)) throw new ArgumentException("message must not be null or empty", nameof(message));
-
             try
             {
                 return enumerable.SingleOrDefault(predicate);
@@ -22,11 +24,9 @@ namespace Nuke.Common.Utilities.Collections
             }
         }
 
+        [CanBeNull]
         public static T SingleOrDefaultOrError<T>(this IEnumerable<T> enumerable, string message)
         {
-            if (enumerable == null) throw new ArgumentNullException(nameof(enumerable));
-            if (string.IsNullOrEmpty(message)) throw new ArgumentException("message must not be null or empty", nameof(message));
-
             try
             {
                 return enumerable.SingleOrDefault();